@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-product-chat-layout',
-  standalone: true,
-  imports: [],
-  templateUrl: './product-chat-layout.component.html',
-  styleUrl: './product-chat-layout.component.scss'
+    selector: 'app-product-chat-layout',
+    imports: [],
+    templateUrl: './product-chat-layout.component.html',
+    styleUrl: './product-chat-layout.component.scss'
 })
 export class ProductChatLayoutComponent {
 
